+++ apps/ymir-sdl3/src/app/settings.hpp
@@ -184,7 +184,7 @@ struct Settings {
         util::Observable<ymir::core::config::sys::VideoStandard> videoStandard;
 
         bool emulateSH2Cache;
-        util::Observable<uint32> sh2OverclockFactor;
+        util::Observable<uint32> sh2ClockFactor;
 
         std::filesystem::path internalBackupRAMImagePath;
         bool internalBackupRAMPerGame;