+++ libs/ymir-core/include/ymir/core/configuration.hpp
@@ -57,11 +57,14 @@ struct Configuration {
         /// Enabling this option incurs a small performance penalty and purges all SH-2 caches.
         util::Observable<bool> emulateSH2Cache = false;
 
-        /// @brief SH-2 overclock factor as a percentage (100 = 1.0x speed).
+        /// @brief SH-2 clock factor as a percentage (100 = 1.0x speed).
         ///
-        /// Adjusts the cycle rate of the SH-2 CPUs, which may help reduce internal slowdowns
-        /// and lag in CPU-heavy games.
-        util::Observable<uint32> sh2OverclockFactor = 100;
+        /// Adjusts the cycle rate of the SH-2 CPUs, which may help reduce internal slowdowns and lag in CPU-heavy
+        /// games.
+        /// Decreasing the factor will improve emulation performance but may cause slowdowns.
+        ///
+        /// Changing it either way from 100% may lower compatibility with some games.
+        util::Observable<uint32> sh2ClockFactor = 100;
     } system;
 
     /// @brief RTC configuration