+++ CHANGELOG.md
@@ -27,11 +27,11 @@ Introduced save state file version 13.
     - When a watchpoint is triggered, Ymir will now display the address and size of the access and which watchpoints were triggered.
 - SCSP: Added Threaded SCSP option, bringing a decent performance boost in nearly all situations. (#3, #896; @SternXD)
 - SCU: Optimize DSP operations involving 48-bit registers.
+- SH2: Clock factor option ranging from 10% up to 300%. (#876, #895; @SternXD)
 - SH2: Decode opcode arguments directly in instruction handlers for better performance.
 - SH2: Fetch pairs of instructions using 32-bit accesses, more closely matching the behavior of the real SH7095. Improves overall emulation accuracy and slightly boosts performance on SH2-bound games.
 - SH2: Handle WB/EX stalls. (thanks to @celeriyacon)
 - SH2: Microoptimize DIV1 instruction.
-- SH2: Overclocking option, up to 300%. (#876, #895; @SternXD)
 
 ### Fixes